# docs/api/openapi.yaml
openapi: 3.0.0
info:
  title: Alfalyzer API
  version: 1.0.0
  description: |
    The Alfalyzer API provides access to financial market data, portfolio management,
    watchlists, and AI-powered earnings transcript analysis.
    
    ## Authentication
    Most endpoints require authentication via Bearer token in the Authorization header.
    
    ## Rate Limiting
    - Anonymous: 60 requests per minute
    - Authenticated: 300 requests per minute
    - Financial data endpoints: 10 requests per minute
    
  contact:
    email: api@alfalyzer.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.alfalyzer.com/v1
    description: Production server
  - url: http://localhost:3001/api
    description: Development server
...
paths:
...
  /stocks/search:
    get:
      tags:
        - Stocks
      summary: Search stocks by symbol or name
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
            minLength: 1
          description: Search query
          example: AAPL
      responses:
        '200':
          description: Search results
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/StockSearchResult'
        '400':
          $ref: '#/components/responses/BadRequest'
  /stocks/{symbol}:
    get:
      tags:
        - Stocks
      summary: Get stock details
      parameters:
        - name: symbol
          in: path
          required: true
          schema:
            type: string
          description: Stock symbol
          example: AAPL
      responses:
        '200':
          description: Stock details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StockDetails'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /watchlists:
    get:
      tags:
        - Watchlists
      summary: List user watchlists
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of watchlists
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Watchlist'
        '401':
          $ref: '#/components/responses/Unauthorized'
    
    post:
      tags:
        - Watchlists
      summary: Create new watchlist
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
              properties:
                name:
                  type: string
                  example: Tech Stocks
                description:
                  type: string
                  example: Technology companies to watch
      responses:
        '201':
          description: Watchlist created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Watchlist'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /watchlists/{id}/stocks:
    post:
      tags:
        - Watchlists
      summary: Add stock to watchlist
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - symbol
              properties:
                symbol:
                  type: string
                  example: AAPL
      responses:
        '201':
          description: Stock added to watchlist
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /transcripts:
    get:
      tags:
        - Transcripts
      summary: List earnings transcripts
      parameters:
        - name: ticker
          in: query
          schema:
            type: string
          description: Filter by stock ticker
        - name: quarter
          in: query
          schema:
            type: string
            enum: [Q1, Q2, Q3, Q4]
        - name: year
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, published]
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
            maximum: 100
      responses:
        '200':
          description: List of transcripts
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/TranscriptSummary'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
...
  /admin/transcripts:
    post:
      tags:
        - Admin
      summary: Upload new transcript
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - ticker
                - company_name
                - quarter
                - year
                - raw_transcript
              properties:
                ticker:
                  type: string
                  example: AAPL
                company_name:
                  type: string
                  example: Apple Inc.
                quarter:
                  type: string
                  enum: [Q1, Q2, Q3, Q4]
                year:
                  type: integer
                  example: 2024
                call_date:
                  type: string
                  format: date
                raw_transcript:
                  type: string
      responses:
        '201':
          description: Transcript created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TranscriptDetails'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    StockSearchResult:
      type: object
      properties:
        symbol:
          type: string
          example: AAPL
        name:
          type: string
          example: Apple Inc.
        type:
          type: string
          example: Common Stock
        exchange:
          type: string
          example: NASDAQ
    StockDetails:
      type: object
      properties:
        symbol:
          type: string
        name:
          type: string
        exchange:
          type: string
        sector:
          type: string
        industry:
          type: string
        marketCap:
          type: number
        pe:
          type: number
        dividendYield:
          type: number
        week52High:
          type: number
        week52Low:
          type: number
        description:
          type: string
...
    Watchlist:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        stockCount:
          type: integer
        stocks:
          type: array
          items:
            type: object
            properties:
              symbol:
                type: string
              name:
                type: string
              price:
                type: number
              change:
                type: number
              changePercent:
                type: number
              addedAt:
                type: string
                format: date-time
    TranscriptSummary:
      type: object
      properties:
        id:
          type: integer
        ticker:
          type: string
        company_name:
          type: string
        quarter:
          type: string
        year:
          type: integer
        call_date:
          type: string
          format: date
        status:
          type: string
          enum: [pending, published]
        has_ai_summary:
          type: boolean
        view_count:
          type: integer
        published_at:
          type: string
          format: date-time
    TranscriptDetails:
      allOf:
        - $ref: '#/components/schemas/TranscriptSummary'
        - type: object
          properties:
            raw_transcript:
              type: string
            ai_summary:
              type: object
              properties:
                summary:
                  type: string
                highlights:
                  type: array
                  items:
                    type: string
                sentiment:
                  type: string
                  enum: [positive, neutral, negative]
    Pagination:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        totalPages:
          type: integer
    Error:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
        statusCode:
          type: integer
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...